silence compiler warnings

+++ lib/lepton/src/CompiledExpression.cpp
@@ -88,7 +88,7 @@ void CompiledExpression::compileExpression(const ExpressionTreeNode& node, vecto
     // Process the child nodes.
 
     vector<int> args;
-    for (int i = 0; i < node.getChildren().size(); i++) {
+    for (int i = 0; i < (int)node.getChildren().size(); i++) {
         compileExpression(node.getChildren()[i], temps);
         args.push_back(findTempIndex(node.getChildren()[i], temps));
     }
@@ -110,7 +110,7 @@ void CompiledExpression::compileExpression(const ExpressionTreeNode& node, vecto
             // If the arguments are sequential, we can just pass a pointer to the first one.
 
             bool sequential = true;
-            for (int i = 1; i < args.size(); i++)
+            for (int i = 1; i < (int)args.size(); i++)
                 if (args[i] != args[i-1]+1)
                     sequential = false;
             if (sequential)
@@ -167,17 +167,17 @@ double CompiledExpression::evaluate() const {
 #ifdef LEPTON_USE_JIT
     return jitCode();
 #else
-    for (int i = 0; i < variablesToCopy.size(); i++)
+    for (int i = 0; i < (int)variablesToCopy.size(); i++)
         *variablesToCopy[i].first = *variablesToCopy[i].second;
 
     // Loop over the operations and evaluate each one.
 
-    for (int step = 0; step < operation.size(); step++) {
+    for (int step = 0; step < (int)operation.size(); step++) {
         const vector<int>& args = arguments[step];
         if (args.size() == 1)
             workspace[target[step]] = operation[step]->evaluate(&workspace[args[0]], dummyVariables);
         else {
-            for (int i = 0; i < args.size(); i++)
+          for (int i = 0; i < (int)args.size(); i++)
                 argValues[i] = workspace[args[i]];
             workspace[target[step]] = operation[step]->evaluate(&argValues[0], dummyVariables);
         }
